Fixed bug where loading a script would remove camera connections
