+++ src/main/java/nl/tudelft/contextproject/ContextTFP.java
@@ -16,10 +16,6 @@ import javafx.scene.image.Image;
 import javafx.scene.layout.BorderPane;
 import javafx.stage.Stage;
 
-import nl.tudelft.contextproject.camera.Camera;
-import nl.tudelft.contextproject.camera.CameraConnection;
-import nl.tudelft.contextproject.camera.LiveCameraConnection;
-import nl.tudelft.contextproject.camera.MockedCameraConnection;
 import nl.tudelft.contextproject.gui.AlertDialog;
 import nl.tudelft.contextproject.gui.MenuController;
 import nl.tudelft.contextproject.saveLoad.ApplicationSettings;
@@ -67,7 +63,6 @@ public class ContextTFP extends Application {
         initRootLayout();
 
         new Thread(() -> initVLCj()).start();
-        new Thread(() -> initCameraConnections()).start();
 
         MenuController.show();
     }
@@ -96,31 +91,6 @@ public class ContextTFP extends Application {
             });
         } catch (IOException e) {
             e.printStackTrace();
-        }
-    }
-
-    /**
-     * Initializes the camera connections for every loaded camera.
-     * If an IP was loaded for a camera, then it will check if it
-     * can make a connection to this camera. If it can, it will
-     * set its connection to a LiveCameraConnection. If it cannot,
-     * it sets a MockedCameraConnection. 
-     */
-    public void initCameraConnections() {
-        ApplicationSettings settings = ApplicationSettings.getInstance();
-
-        for (Camera cam : Camera.getAllCameras()) {
-            String camIp = settings.getCameraIP(cam.getNumber());
-            if (camIp != null && !camIp.equals("")) {
-                CameraConnection connect = new LiveCameraConnection(camIp);
-
-                if (connect.setUpConnection()) {
-                    cam.setConnection(connect);
-                    break;
-                }
-            }
-
-            cam.setConnection(new MockedCameraConnection());
         }
     }
